+++ docs/source/reference/markers/class_scoped_loop_custom_policies_strict_mode_example.py
@@ -3,12 +3,16 @@ import asyncio
 import pytest
 
 
-@pytest.mark.asyncio_event_loop(
-    policy=[
+@pytest.fixture(
+    params=[
         asyncio.DefaultEventLoopPolicy(),
         asyncio.DefaultEventLoopPolicy(),
     ]
 )
+def event_loop_policy(request):
+    return request.param
+
+
 class TestWithDifferentLoopPolicies:
     @pytest.mark.asyncio
     async def test_parametrized_loop(self):